[metasrv] refactor: move sled related files into metasrv/sled_store/
