+++ metasrv/src/meta_service/mod.rs
@@ -12,28 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod applied_state;
-pub mod cmd;
-pub mod errors;
-pub mod log_entry;
-pub mod meta_service_impl;
-pub mod network;
-pub mod placement;
-pub mod raft_db;
-pub mod raft_log;
-pub mod raft_state;
-pub mod raft_state_kv;
-pub mod raft_txid;
-pub mod raft_types;
-pub mod raftmeta;
-pub mod seq_num;
-pub mod sled_key_space;
-pub mod sled_serde;
-pub mod sled_tree;
-pub mod snapshot;
-pub mod state_machine;
-pub mod state_machine_meta;
-
 pub use applied_state::AppliedState;
 pub use cmd::Cmd;
 pub use errors::RetryableError;
@@ -50,12 +28,6 @@ pub use raft_types::NodeId;
 pub use raft_types::Term;
 pub use raftmeta::MetaNode;
 pub use raftmeta::MetaRaftStore;
-pub use seq_num::SeqNum;
-pub use sled_serde::SledOrderedSerde;
-pub use sled_serde::SledSerde;
-pub use sled_tree::AsKeySpace;
-pub use sled_tree::SledTree;
-pub use sled_tree::SledValueToKey;
 pub use snapshot::Snapshot;
 pub use state_machine::Node;
 pub use state_machine::Slot;
@@ -69,6 +41,24 @@ pub use crate::protobuf::meta_service_server::MetaServiceServer;
 pub use crate::protobuf::GetReply;
 pub use crate::protobuf::GetReq;
 pub use crate::protobuf::RaftMes;
+
+pub mod applied_state;
+pub mod cmd;
+pub mod errors;
+pub mod log_entry;
+pub mod meta_service_impl;
+pub mod network;
+pub mod placement;
+pub mod raft_db;
+pub mod raft_log;
+pub mod raft_state;
+pub mod raft_state_kv;
+pub mod raft_txid;
+pub mod raft_types;
+pub mod raftmeta;
+pub mod snapshot;
+pub mod state_machine;
+pub mod state_machine_meta;
 
 #[cfg(test)]
 mod meta_service_impl_test;
@@ -84,10 +74,6 @@ mod raft_state_test;
 mod raft_types_test;
 #[cfg(test)]
 mod raftmeta_test;
-#[cfg(test)]
-mod sled_serde_test;
-#[cfg(test)]
-mod sled_tree_test;
 #[cfg(test)]
 mod state_machine_test;
 #[cfg(test)]
