+++ metasrv/src/meta_service/raft_log.rs
@@ -18,13 +18,13 @@ use async_raft::raft::Entry;
 use common_tracing::tracing;
 
 use crate::configs;
-use crate::meta_service::sled_key_space;
-use crate::meta_service::AsKeySpace;
 use crate::meta_service::LogEntry;
 use crate::meta_service::LogIndex;
-use crate::meta_service::SledSerde;
-use crate::meta_service::SledTree;
-use crate::meta_service::SledValueToKey;
+use crate::sled_store::sled_key_space;
+use crate::sled_store::AsKeySpace;
+use crate::sled_store::SledSerde;
+use crate::sled_store::SledTree;
+use crate::sled_store::SledValueToKey;
 
 const TREE_RAFT_LOG: &str = "raft_log";
