+++ metasrv/src/meta_service/raft_state.rs
@@ -17,13 +17,13 @@ use common_exception::ErrorCode;
 use common_tracing::tracing;
 
 use crate::configs;
-use crate::meta_service::sled_key_space::RaftStateKV;
-use crate::meta_service::AsKeySpace;
 use crate::meta_service::NodeId;
 use crate::meta_service::RaftStateKey;
 use crate::meta_service::RaftStateValue;
-use crate::meta_service::SledSerde;
-use crate::meta_service::SledTree;
+use crate::sled_store::sled_key_space::RaftStateKV;
+use crate::sled_store::AsKeySpace;
+use crate::sled_store::SledSerde;
+use crate::sled_store::SledTree;
 
 /// Raft state stores everything else other than log and state machine, which includes:
 /// id: NodeId,
